requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.28
redis==4.3.4
werkzeug==2.2.3
